error_injection: Reinvent exception injection
There's .inject() overload that accepts exception factory, but it's not
used at all (test doesn't count). At the same time all error injections
that want to inject exception use the

  get_local_injector().inject(name, [] { throw some_exception(reason); });

format. So it makes sense to reinvent the exception injection overload
to accept class name and thrown it.

One note here -- typically `reason` matches the `name` or is derived
from it, so shorten the API here as well and throw exception with the
injection name, not dedicated reason.

Signed-off-by: Pavel Emelyanov <xemul@scylladb.com>