sstables_loader: Stream to pending tablet replica if needed
Even though taking erm blocks migration, it cannot prevent the
load-and-stream to start while a migration is going on, erm
only prevents migration from advancing.

With tablets, new data will be streamed to pending replica too if
the write replica selector, in transition metadata, is set to both.
If migration is at a later stage where only new replica is written
to, then data is streamed only to new replica as selector is set
to next (== new replica set).

primary_replica_only flag is handled by only streaming to pending
if the primary replica is the one leaving through migration.

Signed-off-by: Raphael S. Carvalho <raphaelsc@scylladb.com>