tools/scylla-nodetool: add {json,yaml}_writer
so that we have less repeatings for dumping the metrics. the repeatings
are error-prone and not maintainable. also move them out into a separate
header, to keep fit of this source file -- it's now 3000 LOC. also,
by moving them out, we can reuse them in other subcommands without
moving them to the top of this source file.

Signed-off-by: Kefu Chai <kefu.chai@scylladb.com>