test/pylib: merge default option and customized option
before this change, the options specified by the `create_server`
callback always override the ones returned by `make_scylla_conf()`,
and the result is used for creating a server instance for testing.
and the former is always provided by per-suite's `suite.yaml` option.
they are well defined so that the options are sane. but if we allow
user to enable extra experimental features, they should not override
the ones set by `make_scylla_conf()`. otherwise the tester who
adds the extra experimental features would have to adjust the
settings to find a sane combination.

so, in this change, instead overriding the options with the value
provided by `create_server` callback, let's adding them if the value
is a list.

Refs #17406
Signed-off-by: Kefu Chai <kefu.chai@scylladb.com>