compaction: handle nullptr owned_ranges in get_ranges_for_invalidation
Commit 7a988777 sets owned_ranges_ptr to null when initiating an upgrade
on keyspaces with tablets. owned_ranges_ptr is set to null because
cleanup is not required for tables using tablets. As a result, cache
invalidation is also not required for any ranges. Update
get_ranges_for_invalidation to handle null values of owned_ranges_ptr by
returning empty ranges for cache invalidation.

Fixes #17452

Signed-off-by: Lakshmi Narayanan Sreethar <lakshmi.sreethar@scylladb.com>