test.py: Add test-case splitting in 'name' selection
When filtering a test by 'name' consider that name can be in a
'test::case' format. If so, get the left part to be the filter and the
right part to be the case name to be passed down to test itself.

Later, when the pytest starts it then appends the case name (if not
None) to the pytest execution, thus making it run only the specified
test-case, not the whole test file.

Signed-off-by: Pavel Emelyanov <xemul@scylladb.com>