test: topology: increase max reconnection retries in cql driver
Previous max_retries default value was 64 which means we gave up after
~4s * 64 = ~4mins. But we have TOPOLOGY_TIMEOUT = 1000 (seconds)
which is a bit larger.

So now we set new default limit to around 34 mins.

Some indicator that it may help with failing tests is that
I observed 'Will not continue to retry reconnection attempts
due to an exhausted retry schedule' warning in the logs of
failed test in debug mode.